Rename framework to ValidatorKit
This is because ReactiveCocoa introduces a ValidationResult type and Xcode couldn’t figure out the difference between the two.

We also couldn't use the Module.Class as the Validator module has the same class as the name.
Things that didn't work:
ValidationResult
Validator.ValidationResult
Validator.Validator.ValidationReult

Due to the module name being the same as the Validator class name. Solution is this, which renames the module. This will allow us to specify exactly which file (and type) we want to use in our project.